+++ proofs/application.php
@@ -48,12 +48,11 @@ return static function() {
                 ->scenariiPerProof($scenarii)
                 ->displayOutputVia($io)
                 ->displayErrorVia($io)
-                ->tryToProve(static function() {
-                    yield proof(
-                        'example',
-                        given(Set\Integers::any()),
-                        static fn($assert, $i) => $assert->true(true),
-                    );
+                ->tryToProve(static function($prove) {
+                    yield $prove
+                        ->proof('example')
+                        ->given(Set\Integers::any())
+                        ->test(static fn($assert, $i) => $assert->true(true));
                 });
 
             $assert->true($result->successful());
@@ -73,8 +72,8 @@ return static function() {
                 ->scenariiPerProof($scenarii)
                 ->displayOutputVia($io)
                 ->displayErrorVia($io)
-                ->tryToProve(static function() {
-                    yield property(
+                ->tryToProve(static function($prove) {
+                    yield $prove->property(
                         LowerBoundAtZero::class,
                         Set\Elements::of(new Counter),
                     );
@@ -98,8 +97,8 @@ return static function() {
                 ->displayOutputVia($io)
                 ->displayErrorVia($io)
                 ->allowProofsToNotMakeAnyAssertions()
-                ->tryToProve(static function() {
-                    yield properties(
+                ->tryToProve(static function($prove) {
+                    yield $prove->properties(
                         'Counter properties',
                         Set\Properties::any(
                             DownAndUpIsAnIdentityFunction::any(),